UserMenu done and UserGroupMenu done = every function works.
Adjusted css classes in these components.

Changed most Contracts( and Classes, Services) to work with Lists, not IEnums or Collections.

Some contracts now include whole objects instead of IDs. This should be useful in the future. It should have close to 0 impact on the amount of data transfered during requests.

TODO:
Moving roles in UserMenu refuses to animate.